+++ grid2op/Converter/__init__.py
@@ -12,11 +12,3 @@ from grid2op.Converter.ToVect import ToVect
 from grid2op.Converter.IdToAct import IdToAct
 from grid2op.Converter.AnalogStateConverter import AnalogStateConverter
 from grid2op.Converter.ConnectivityConverter import ConnectivityConverter
-
-try:
-    from grid2op.Converter.GymConverter import GymObservationSpace, GymActionSpace
-    __all__.append("GymObservationSpace")
-    __all__.append("GymActionSpace")
-except ImportError:
-    # you must install open ai gym to benefit from this converter
-    pass
