+++ grid2op/gym/__init__.py
@@ -1,0 +1,25 @@
+__all__ = ["BaseGymAttrConverter", "GymEnv"]
+from grid2op.gym.base_gym_attr_onverter import BaseGymAttrConverter
+from grid2op.gym.gymenv import GymEnv
+
+try:
+    from grid2op.gym.gym_space_converter import GymObservationSpace, GymActionSpace
+    __all__.append("GymObservationSpace")
+    __all__.append("GymActionSpace")
+except ImportError:
+    # you must install open ai gym to benefit from this converter
+    pass
+
+try:
+    from grid2op.gym.scaler_attr_converter import ScalerAttrConverter
+    __all__.append("ScalerAttrConverter")
+except ImportError:
+    # you must install open ai gym to benefit from this converter
+    pass
+
+try:
+    from grid2op.gym.multi_to_tuple_converter import MultiToTupleConverter
+    __all__.append("MultiToTupleConverter")
+except ImportError:
+    # you must install open ai gym to benefit from this converter
+    pass
